+++ tests/backend-api/abclassMutations.unit.test.js
@@ -5,7 +5,7 @@
 
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
-const modulePath = '../../src/backend/z_Api/abclassMutations.js';
+const modulePath = '../../src/backend/z_Api/abclass/abclassMutations.js';
 const ApiValidationError = require('../../src/backend/Utils/ErrorTypes/ApiValidationError.js');
 
 let originalABClassController;
